Added PEST Template File Creation
Added capability to create a PEST template file for the SFR package.  The variable SFR_Template was added to the general_inputs.txt file, which can either be PEST or UCODE and determines which template file type (.jtf for UCODE or .tpl for PEST) is generated by the SWBM.

Removed unnamed files (e.g., fort.611) that were generated for testing input/output and no longer necessary.

Added printing of irrigation efficiencies to the SWBM record file for bookkeeping purposes.