Update of SWBM Code
Changed how MAR is passed from SWBM to MODFLOW. MAR was previously added directly to recharge matrix, bypassing the soil-zone completely. MAR is now added as additional irrigation from Jan-Mar. Fields are now specified with a separate input file (MAR_Fields.txt) that gives the total number of fields followed by the field ID and teh maximum infiltration rate. The model then allocates the total amount of MAR available to the fields in the order they appear in the file. If there is a limited amount of MAR, fields near the end will not receive any.

Also updated the code for daily output on specific fields. User can now specify which fields have daily budgets printed for them in the daily_out.txt file. There is also a flag for turning off printing of daily budgets.