Updated SWBM
Corrected mismatch between SWBM output and MODFLOW inputs. The discrepency between the actual field area and the field area in the model due to the mapping of the fields onto the MODFLOW grid was resulting in differing values in the water budget figures. The SWBM now uses the MODFLOW field area for all of the calculations instead of the actual field area. This removes the fields smaller than the MODFLOW grid cell size, but overall only changes the total field area by about 5.5% and produces a similar water budget by landuse type.